Replacing the "grid" color schema for a pixel shader.
Now the influence is calculated also by the shader and the color of each graivity influence is produced by the shaders.

The grid still being printed (with the force values) just for troubleshooting purposes.

Eventually the "resulting force" will onlly be calcaulated for "game objects" instead of a entire grid